requirements.txt
opencv_python==4.5.5.62
Pillow==9.1.0
